Fxing bug when EJSApp booking and/or Remlab Manager plugins are not installed

+++ lib.php
@@ -175,21 +175,29 @@ function ejsapp_update_instance($ejsapp, $mform=null) {
     $fs->delete_area_files($context->id, 'mod_ejsapp', 'tmp_jarfiles', $ejsapp->id);
     $ejs_ok = update_ejsapp_files_and_tables($ejsapp, $context);
     if ($ejs_ok) {
+        $is_ejsappbooking_installed = $DB->get_records('modules',array('name'=>'ejsappbooking'));
+        $is_ejsappbooking_installed = !empty($is_ejsappbooking_installed);
+        $is_remlab_manager_installed = $DB->get_records('modules',array('name'=>'remlab_manager'));
+        $is_remlab_manager_installed = !empty($is_remlab_manager_installed);
         if ($ejsapp->is_rem_lab == 1) { // Remote lab
-            $complete_pract_list = explode(';', $ejsapp->list_practices);
-            $remlab_info = $DB->get_record('remlab_manager_conf', array('practiceintro' => $complete_pract_list[$ejsapp->practiceintro]));
-            if ($remlab_info == null) {
-                $remlab_info = default_rem_lab_conf($ejsapp);
-                $DB->insert_record('remlab_manager_conf', $remlab_info);
+            if ($is_remlab_manager_installed) {
+                $complete_pract_list = explode(';', $ejsapp->list_practices);
+                $remlab_info = $DB->get_record('remlab_manager_conf', array('practiceintro' => $complete_pract_list[$ejsapp->practiceintro]));
+                if ($remlab_info == null) {
+                    $remlab_info = default_rem_lab_conf($ejsapp);
+                    $DB->insert_record('remlab_manager_conf', $remlab_info);
+                }
+                $DB->delete_records('remlab_manager_expsyst2pract', array('ejsappid' => $ejsapp->id));
+                ejsapp_expsyst2pract($ejsapp);
             }
-            $DB->delete_records('remlab_manager_expsyst2pract', array('ejsappid' => $ejsapp->id));
-            ejsapp_expsyst2pract($ejsapp);
-            update_booking_table($ejsapp);
+            if ($is_ejsappbooking_installed) update_booking_table($ejsapp);
         } else {
-            $DB->delete_records('remlab_manager_expsyst2pract', array('ejsappid' => $ejsapp->id));
-            if($DB->record_exists('ejsappbooking', array('course'=>$ejsapp->course))) {
-              $DB->delete_records('ejsappbooking_usersaccess', array('ejsappid' => $ejsapp->id));
-              $DB->delete_records('ejsappbooking_remlab_access', array('ejsappid' => $ejsapp->id));
+            if ($is_remlab_manager_installed) $DB->delete_records('remlab_manager_expsyst2pract', array('ejsappid' => $ejsapp->id));
+            if ($is_ejsappbooking_installed) {
+                if ($DB->record_exists('ejsappbooking', array('course' => $ejsapp->course))) {
+                    $DB->delete_records('ejsappbooking_usersaccess', array('ejsappid' => $ejsapp->id));
+                    $DB->delete_records('ejsappbooking_remlab_access', array('ejsappid' => $ejsapp->id));
+                }
             }
         }
     } else {
